saas-swarm/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
numpy>=1.26.0
pydantic>=2.5.0
pyyaml>=6.0.1
click>=8.1.7
asyncio-mqtt>=0.16.1
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0 